+++ src/api.jl
@@ -244,8 +244,15 @@ function ts_parser_language(self)
     ccall((:ts_parser_language, libtreesitter), Ptr{TSLanguage}, (Ptr{TSParser},), self)
 end
 
-function ts_parser_set_included_ranges()
-    ccall((:ts_parser_set_included_ranges, libtreesitter), Cint, ())
+function ts_parser_set_included_ranges(self, ranges, count)
+    ccall(
+        (:ts_parser_set_included_ranges, libtreesitter),
+        Bool,
+        (Ptr{TSParser}, Ptr{TSRange}, UInt32),
+        self,
+        ranges,
+        count,
+    )
 end
 
 function ts_parser_included_ranges(self, length)
@@ -296,34 +303,6 @@ end
 
 function ts_parser_reset(self)
     ccall((:ts_parser_reset, libtreesitter), Cvoid, (Ptr{TSParser},), self)
-end
-
-function ts_parser_set_timeout_micros(self, timeout)
-    ccall(
-        (:ts_parser_set_timeout_micros, libtreesitter),
-        Cvoid,
-        (Ptr{TSParser}, UInt64),
-        self,
-        timeout,
-    )
-end
-
-function ts_parser_timeout_micros(self)
-    ccall((:ts_parser_timeout_micros, libtreesitter), UInt64, (Ptr{TSParser},), self)
-end
-
-function ts_parser_set_cancellation_flag(self, flag)
-    ccall(
-        (:ts_parser_set_cancellation_flag, libtreesitter),
-        Cvoid,
-        (Ptr{TSParser}, Ptr{Cint}),
-        self,
-        flag,
-    )
-end
-
-function ts_parser_cancellation_flag()
-    ccall((:ts_parser_cancellation_flag, libtreesitter), Ptr{Cint}, ())
 end
 
 function ts_parser_set_logger(self, logger)